new `hash_file_path` is a superset in functionality compared to old
